refactor(security): extract curl probe policy
